requirements.txt
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
civil_tools_v
contourpy==1.3.1
cycler==0.12.1
daphne==4.1.2
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
ezdxf==1.4.0
fonttools==4.56.0
kiwisolver==1.4.8
lxml==5.3.1
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1